update users_serviec facade moving exceptions to new leaf module error
