+++ services/web/server/src/simcore_service_webserver/announcements/announcements_service.py
@@ -1,4 +1,3 @@
-# Functions
 from ._api import list_announcements
 
 __all__: tuple[str, ...] = (
