+++ services/web/server/src/simcore_service_webserver/user_preferences/_service.py
@@ -107,7 +107,7 @@ async def set_frontend_user_preference(
     frontend_preference_identifier: PreferenceIdentifier,
     value: Any,
 ) -> None:
-    from ..users.exceptions import FrontendUserPreferenceIsNotDefinedError  # noqa: PLC0415
+    from ..users.errors import FrontendUserPreferenceIsNotDefinedError  # noqa: PLC0415
 
     try:
         preference_name: PreferenceName = get_preference_name(frontend_preference_identifier)
