+++ services/web/server/src/simcore_service_webserver/users/_controller/rest/users_rest.py
@@ -16,10 +16,10 @@ from servicelib.aiohttp.requests_validation import (
 from simcore_service_webserver.application_settings_utils import (
     requires_dev_feature_enabled,
 )
-from simcore_service_webserver.users.exceptions import BillingDetailsNotFoundError
+from simcore_service_webserver.users.errors import BillingDetailsNotFoundError
 
 from ...._meta import API_VTAG
-from ....groups.groups_service import get_user_profile_groups
+from ....groups import groups_service
 from ....login.decorators import login_required
 from ....products import products_web
 from ....products.models import Product
@@ -56,7 +56,7 @@ async def get_my_profile(request: web.Request) -> web.Response:
         my_product_group,
         product_support_group,
         product_chatbot_primary_group,
-    ) = await get_user_profile_groups(request.app, user_id=req_ctx.user_id, product=product)
+    ) = await groups_service.get_user_profile_groups(request.app, user_id=req_ctx.user_id, product=product)
 
     assert groups_by_type.primary  # nosec
     assert groups_by_type.everyone  # nosec
