+++ services/web/server/src/simcore_service_webserver/users/_users_service.py
@@ -28,7 +28,7 @@ from ._models import (
     UserIdNamesTuple,
     UserModelAdapter,
 )
-from .exceptions import (
+from .errors import (
     MissingGroupExtraPropertiesForProductError,
     UserNotFoundError,
 )
